Repair file casing
