+++ client/src/components/Home/Welcome/welcome.scss
@@ -1,5 +1,5 @@
 .welcome{
-  background: url('http://res.cloudinary.com/otse/image/upload/q_auto:low/v1531698049/crispy-fish_ztgkfr.jpg');
+  background: url('//res.cloudinary.com/otse/image/upload/q_auto:low/v1531698049/crispy-fish_ztgkfr.jpg');
   background-repeat: no-repeat;
   background-size: cover;
 	width: 100%;
