+++ client/src/components/OrderModal/order-modal.scss
@@ -181,6 +181,7 @@
   cursor: pointer;
   outline: none;
 }
+
 .successBtn {
   @extend .btn;
   color: white;
@@ -189,6 +190,14 @@
   &:hover{
     background-color: darken($green-btn-color, 5%);
     box-shadow: 0 1px 3px rgba(0, 0, 0, 0.12), 0 1px 2px rgba(0, 0, 0, 0.24);
+  }
+
+  &:disabled, &[disabled]{
+    background-color: darken($green-btn-color, 5%);
+    &:hover{
+      box-shadow: none;
+      cursor: none;
+    }
   }
 }
