+++ client/src/components/Orders/order.scss
@@ -126,6 +126,7 @@
 }
 
 .deliveryAddress {
+  margin-top: 0;
   @media only screen and (min-width: 1200px) {
     max-width: 307px;
   }
